demo/requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
blobfile==3.0.0
cachetools==6.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
docker-pycreds==0.4.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-gpu==1.7.2
fastapi==0.116.1
filelock==3.13.1
fonttools==4.59.0
frozenlist==1.6.0
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf_transfer==0.1.9
huggingface-hub==0.31.2
idna==3.10
ipython==8.37.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
joblib==1.5.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kagglehub==0.3.12
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.4.0
MarkupSafe==2.1.5
matplotlib==3.10.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
narwhals==2.1.2
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
open_clip_torch==2.32.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pycryptodomex==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-stubs==5.15.6.0
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.27.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.28.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
streamlit==1.48.1
sympy==1.13.3
tenacity==9.1.2
tensorboard-logger==0.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.3
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0+cu118
torchao==0.11.0
torchaudio==2.7.0+cu118
torchdata==0.11.0
torchtune==0.6.1
torchvision==0.22.0+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
wandb==0.19.11
watchdog==6.0.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
